Query autonomy profile by given coordinates (#97)
In this PR, we add the support of querying autonomy profile by given a random coordinate. 

When a user searches for a location, there will be profile page for them before they really add it. The place might not be added into POI before. By searching coordinates, we will be able to address whether it is a existent POI or not. If it is not, we will calculate a score immediate based on the given coordinates.

Changes Highlight:
1. Add a DB function `GetPOIByCoordinates` to search POI by coordinates
2. Create a help function `summarizePlaceProfile` to generate response body for autonomy profile of a POI.
3. Slightly refactoring for the `placeProfile` handler.